backend/requirements.txt
flask
Flask-RESTful
python-dotenv
psycopg2-binary